+++ CHANGES.rst
@@ -5,7 +5,14 @@ Changelog of nens-graph
 0.1 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Changed signature of NensGraph.on_draw()
+
+- Added RainappGraph
+
+- Added pixel-to-figure-coordinates methods for NensGraph class
+
+- Added methods to find out width and height of specific plot objects at drawing
+  time
 
 
 0.0.1 (2011-05-18)